fix Task espresso tests, refactor package structure to move DI stuff into DI folder
