+++ todoapp/app/src/main/java/com/example/android/architecture/blueprints/todoapp/di/PerActivity.java
@@ -1,4 +1,4 @@
-package com.example.android.architecture.blueprints.todoapp.util;
+package com.example.android.architecture.blueprints.todoapp.di;
 
 import com.example.android.architecture.blueprints.todoapp.di.AppComponent;
