+++ todoapp/app/src/main/java/com/example/android/architecture/blueprints/todoapp/statistics/StatisticsModule.java
@@ -1,7 +1,7 @@
 package com.example.android.architecture.blueprints.todoapp.statistics;
 
-import com.example.android.architecture.blueprints.todoapp.util.PerActivity;
-import com.example.android.architecture.blueprints.todoapp.util.PerFragment;
+import com.example.android.architecture.blueprints.todoapp.di.PerActivity;
+import com.example.android.architecture.blueprints.todoapp.di.PerFragment;
 
 import dagger.Binds;
 import dagger.Module;
