+++ todoapp/app/src/main/java/com/example/android/architecture/blueprints/todoapp/tasks/TasksModule.java
@@ -1,11 +1,10 @@
 package com.example.android.architecture.blueprints.todoapp.tasks;
 
-import com.example.android.architecture.blueprints.todoapp.util.PerActivity;
-import com.example.android.architecture.blueprints.todoapp.util.PerFragment;
+import com.example.android.architecture.blueprints.todoapp.di.PerActivity;
+import com.example.android.architecture.blueprints.todoapp.di.PerFragment;
 
 import dagger.Binds;
 import dagger.Module;
-import dagger.Provides;
 import dagger.android.ContributesAndroidInjector;
 
 /**
